Moves DiscoveryUtil to discovery package.
