+++ src/main/java/org/jitsi/impl/protocol/xmpp/OpSetSimpleCapsImpl.java
@@ -19,7 +19,7 @@ package org.jitsi.impl.protocol.xmpp;
 
 import net.java.sip.communicator.util.*;
 
-import org.jitsi.jicofo.util.*;
+import org.jitsi.jicofo.discovery.*;
 import org.jitsi.protocol.xmpp.*;
 
 import org.jivesoftware.smack.*;
