+++ src/main/java/org/jitsi/jicofo/discovery/DiscoveryUtil.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jitsi.jicofo.util;
+package org.jitsi.jicofo.discovery;
 
 import net.java.sip.communicator.impl.protocol.jabber.extensions.health.*;
 import net.java.sip.communicator.service.protocol.*;
