+++ src/test/java/mock/muc/MockRoomMember.java
@@ -18,9 +18,11 @@
 package mock.muc;
 
 import mock.xmpp.*;
+
 import net.java.sip.communicator.service.protocol.*;
 import net.java.sip.communicator.service.protocol.globalstatus.*;
-import org.jitsi.jicofo.util.*;
+
+import org.jitsi.jicofo.discovery.*;
 import org.jitsi.protocol.xmpp.*;
 
 import java.util.*;
